src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/scrollbar-expected.txt

CONSOLE ERROR: line 34: Uncaught TypeError: host.createShadowRoot is not a function
